Throw exception if state-modifying HookM code passed between components (#44)
Throw exception if state-modifying `HookM` code is passed between components

Defining code to run in `HookM` is simply defining a sequence of instructions. These
instructions are then executed in the context of a particular component -- the
component is the "interpreter" for the `HookM` code.

Since the component is the interpreter for the `HookM` code it's necessary that code
that interacts with the interpreter's environment (e.g. state-modifying code) is only run
in the interpreter where it was defined. Otherwise you can end up doing things like
modifying the wrong state or attempting to access an index in state that doesn't
exist in the environment. This leads to either subtle bugs or outright crashes without
an informative error message.

This commit ensures that `HookM` code which modifies state throws an exception if
it is created in one component and then executed in another. It is, unfortunately, not
a compile-time error; the usability hit of making this a compile-time error is simply
too large.